requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements-dev.in
#
astroid==2.9.3
    # via pylint
attrs==21.4.0
    # via pytest
black==21.12b0
    # via -r requirements-dev.in
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via pynacl
charset-normalizer==2.0.10
    # via requests
click==8.0.3
    # via
    #   -c requirements.txt
    #   black
coverage[toml]==6.3
    # via pytest-cov
idna==3.3
    # via
    #   -c requirements.txt
    #   requests
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via pylint
lazy-object-proxy==1.7.1
    # via astroid
mccabe==0.6.1
    # via pylint
mypy==0.931
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
platformdirs==2.4.1
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pylint==2.12.2
    # via -r requirements-dev.in
pynacl==1.5.0
    # via -r requirements-dev.in
pyparsing==3.0.6
    # via packaging
pytest==6.2.5
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-mock
pytest-cov==3.0.0
    # via -r requirements-dev.in
pytest-mock==3.6.1
    # via -r requirements-dev.in
requests==2.27.1
    # via -r requirements-dev.in
toml==0.10.2
    # via
    #   -c requirements.txt
    #   pylint
    #   pytest
tomli==1.1.0
    # via
    #   -c requirements.txt
    #   black
    #   coverage
    #   mypy
types-orjson==3.6.2
    # via -r requirements-dev.in
types-pyyaml==6.0.3
    # via -r requirements-dev.in
types-ujson==4.2.1
    # via -r requirements-dev.in
typing-extensions==4.0.1
    # via
    #   -c requirements.txt
    #   astroid
    #   black
    #   mypy
    #   pylint
urllib3==1.26.8
    # via requests
wrapt==1.13.3
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
